requirements.txt
annotated-types==0.7.0
anyio==4.7.0
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
fastapi==0.115.6
h11==0.14.0
idna==3.10
pdfminer.six==20240706
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
